Fix many warnings (#997)
* Fix mismatched-tags warning

* Fix pessimizing-move warnings

* Fix reorder-ctor warnings

* Fix unused-variable warnings

* Fix uninitialized warnings

* Fix sign-compare warnings

* Fix catch-value warnings

* Fix misleading-indentation warning, CHANGES BEHAVIOR

* Hack memcopy to work around class-memaccess warnings

* Fix initialization to default

* Fix cast to char*

---------

+++ src/distribution/AtlasDistribution.cc
@@ -63,7 +63,7 @@ AtlasDistribution::RecordAssigner::RecordAssigner(const Parameters_ & params)
   atlas::MeshGenerator generator(atlasConfig);
 
   mesh_ = generator.generate(grid);
-  if (mesh_->nb_partitions() != atlas::mpi::comm().size()) {
+  if (static_cast<size_t>(mesh_->nb_partitions()) != atlas::mpi::comm().size()) {
     std::stringstream msg;
     msg << "The number of mesh partitions, " << mesh_->nb_partitions()
         << ", is different from the number of MPI processes, " << atlas::mpi::comm().size();
@@ -97,7 +97,7 @@ bool AtlasDistribution::RecordAssigner::isInMyDomain(const eckit::geometry::Poin
   const atlas::idx_t partition = (*locator_)(point);
   oops::Log::debug() << "RecordAssigner::isInMyDomain(): Polygon locator says "
                      << point << " is in domain " << partition << std::endl;
-  return partition == atlas::mpi::comm().rank();
+  return (static_cast<size_t>(partition) == atlas::mpi::comm().rank());
 }
 
 // -----------------------------------------------------------------------------
